+++ src/components/App/AppDumb.js
@@ -6,6 +6,7 @@ import Nav from '../Nav/Nav';
 import Products from '../Products';
 import SignUp from '../SignUp/SignUp';
 import ItemList from '../ItemList/ItemList';
+import Checkout from '../Checkout/Checkout';
 
 class App extends Component {
   // fetching beverages here is temporary, but right spot for cart and user
@@ -23,6 +24,7 @@ class App extends Component {
         <GuestOnly exact={true} path="/signup" component={SignUp} />
         <Route exact path="/products" component={Products} />
         <Route exact path="/cart" component={ItemList} />
+        <Route exact path="/checkout" component={Checkout} />
       </HashRouter>
     );
   }
